backend/requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
python-multipart==0.0.21
python-dotenv==1.2.1
google-genai==1.56.0
supabase==2.27.0
pydantic==2.10.6